Remove BaseBlockService
