+++ Block/Service/EmptyBlockService.php
@@ -12,13 +12,13 @@
 namespace Sonata\BlockBundle\Block\Service;
 
 use Sonata\AdminBundle\Form\FormMapper;
-use Sonata\BlockBundle\Block\BaseBlockService;
+use Sonata\BlockBundle\Block\AbstractAdminBlockService;
 use Sonata\BlockBundle\Block\BlockContextInterface;
 use Sonata\BlockBundle\Model\BlockInterface;
 use Sonata\CoreBundle\Validator\ErrorElement;
 use Symfony\Component\HttpFoundation\Response;
 
-class EmptyBlockService extends BaseBlockService
+class EmptyBlockService extends AbstractAdminBlockService
 {
     /**
      * {@inheritdoc}
